requirements.txt
asgiref==3.2.7
Django==3.0.7
django-crispy-forms==1.9.1
django-filter==2.3.0
djangorestframework==3.11.0
Markdown==3.2.2
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
